refactor search list
