+++ src/routes/List/Applications.less
