+++ src/routes/List/List.js
@@ -1,0 +1,82 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
+import { routerRedux, Route, Switch } from 'dva/router';
+import { connect } from 'dva';
+import { Input } from 'antd';
+import PageHeaderLayout from '../../layouts/PageHeaderLayout';
+
+@connect()
+export default class SearchList extends Component {
+  static contextTypes = {
+    routeData: PropTypes.array,
+  };
+
+  handleTabChange = (key) => {
+    const { dispatch } = this.props;
+    switch (key) {
+      case 'articles':
+        dispatch(routerRedux.push('/list/search/articles'));
+        break;
+      case 'applications':
+        dispatch(routerRedux.push('/list/search/applications'));
+        break;
+      case 'projects':
+        dispatch(routerRedux.push('/list/search/projects'));
+        break;
+      default:
+        break;
+    }
+  }
+
+  render() {
+    const tabList = [{
+      key: 'articles',
+      tab: '文章',
+    }, {
+      key: 'applications',
+      tab: '应用',
+    }, {
+      key: 'projects',
+      tab: '项目',
+    }];
+
+    const mainSearch = (
+      <div style={{ textAlign: 'center' }}>
+        <Input.Search
+          placeholder="请输入"
+          enterButton="搜索"
+          size="large"
+          onSearch={this.handleFormSubmit}
+          style={{ width: 522 }}
+        />
+      </div>
+    );
+
+    const { routeData } = this.context;
+    const routes = routeData.filter(item => item.path === '/list/search')[0].children;
+    const { match } = this.props;
+
+    return (
+      <PageHeaderLayout
+        title="搜索列表"
+        content={mainSearch}
+        tabList={tabList}
+        onTabChange={this.handleTabChange}
+      >
+        <Switch>
+          {
+            routes.map(item =>
+              (
+                <Route
+                  key={item.path}
+                  path={`${match.url}/${item.path}`}
+                  component={item.component}
+                />
+              )
+            )
+          }
+        </Switch>
+      </PageHeaderLayout>
+    );
+  }
+}
